Break apart AuditLog module into separate partial classes
